requirement.txt
python-telegram-bot==13.13
openai==0.10.1
urllib3
six
